+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+5.2.0
+-----
+
+ * deprecated passing of `database` or `collection` to `MongoDbStore` via connection URI, use `$options` instead.
+
 5.1.0
 -----
 
@@ -19,10 +24,10 @@ CHANGELOG
  * added InvalidTtlException
  * deprecated `StoreInterface` in favor of `BlockingStoreInterface` and `PersistingStoreInterface`
  * `Factory` is deprecated, use `LockFactory` instead
- * `StoreFactory::createStore` allows PDO and Zookeeper DSN. 
- * deprecated services `lock.store.flock`, `lock.store.semaphore`, `lock.store.memcached.abstract` and `lock.store.redis.abstract`, 
+ * `StoreFactory::createStore` allows PDO and Zookeeper DSN.
+ * deprecated services `lock.store.flock`, `lock.store.semaphore`, `lock.store.memcached.abstract` and `lock.store.redis.abstract`,
    use `StoreFactory::createStore` instead.
-    
+
 4.2.0
 -----